FitMind/requirements.txt
accelerate==0.15.0
ailment==9.2.39
aiohttp==3.8.3
aiosignal==1.3.1
angr==9.2.39
appdirs==1.4.4
Appium-Python-Client==2.8.1
archinfo==9.2.39
asgiref==3.7.2
astroid==2.14.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
banana-dev==3.0.1
bitstring==4.0.1
cachetools==5.3.0
capstone==4.0.2
cattrs==22.2.0
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==2.1.1
claripy==9.2.39
cle==9.2.39
click==8.1.3
colorama==0.4.5
coverage==6.5.0
CppHeaderParser==2.7.4
crosshair-tool==0.0.37
cryptography==38.0.1
crytic-compile==0.2.2
cuda-python==11.8.1
Cython==0.29.32
cytoolz==0.12.1
deap==1.3.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
Django==4.1.2
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-cors-headers==4.3.1
django-js-asset==2.2.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
docker-pycreds==0.4.0
dpkt==1.9.8
eth-hash==0.5.1
eth-typing==3.2.0
eth-utils==2.1.0
exceptiongroup==1.1.0
execnet==1.9.0
filelock==3.8.0
Flask==2.2.2
frozenlist==1.3.3
future==0.18.3
gitdb==4.0.10
GitPython==3.1.30
graphviz==0.20.1
grpcio==1.51.1
h11==0.14.0
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.12.0
iniconfig==1.1.1
intervaltree==3.1.0
isort==5.12.0
itanium-demangler==1.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lazy-object-proxy==1.9.0
llvmlite==0.40.0
lsprotocol==2022.0.0a10
manticore==0.3.7
MarkupSafe==2.1.1
mccabe==0.7.0
moodlepy==0.24.0
mpi4py==3.1.4
mpmath==1.2.1
msgpack==1.0.7
mulpyplexer==0.9
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.0
nampa==0.1.1
networkx==3.0
nlpcloud==1.0.38
numba==0.57.0
numpy==1.23.5
oauthlib==3.2.2
openai==0.26.4
outcome==1.2.0
packaging==21.3
pathtools==0.1.2
pefile==2023.2.7
Pillow==9.2.0
platformdirs==3.1.0
pluggy==1.0.0
plumbum==1.8.1
ply==3.11
prettytable==3.6.0
progressbar2==4.2.0
protobuf==4.21.12
psutil==5.9.4
py==1.11.0
pycparser==2.21
pyelftools==0.29
pyevmasm==0.2.3
pygame==2.1.2
pygls==1.0.1
PyJWT==1.7.1
pylint==2.16.3
pyparsing==3.0.9
pysha3==1.0.2
PySMT==0.9.5
PySocks==1.7.1
pytest==7.1.3
pytest-cov==4.0.0
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dotenv==0.19.0
python-utils==3.5.2
python3-openid==3.2.0
pytz==2022.5
pyvex==9.2.39
pywin32==305
PyYAML==6.0
readerwriterlock==1.0.9
redis==5.0.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rlp==3.0.0
rpyc==5.3.1
selenium==4.8.2
sentry-sdk==1.15.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
stripe==7.9.0
sympy==1.11.1
tokenizers==0.13.2
tomli==2.0.1
tomlkit==0.11.6
toolz==0.12.0
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
transformers==4.24.0
trio==0.22.0
trio-websocket==0.9.2
typeguard==2.13.3
typeshed-client==2.2.0
typing-inspect==0.8.0
typing_extensions==4.9.0
tzdata==2022.7
unicorn==2.0.1.post1
urllib3==1.26.12
wandb==0.13.10
wasm==1.2
wcwidth==0.2.6
Werkzeug==2.2.2
wrapt==1.15.0
wsproto==1.2.0
yarl==1.8.2
z3-solver==4.11.2.0
zipp==3.11.0
